refactor: 全モジュール（business, admin, user）のID型をint32に標準化

+++ backend/admin/handler/admin_contact_handler.go
@@ -63,7 +63,7 @@ func (h *AdminContactHandler) ApproveInquiry(c *gin.Context) {
 		return
 	}
 
-	err = h.service.ApproveInquiry(id)
+	err = h.service.ApproveInquiry(int32(id))
 	if err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
@@ -90,7 +90,7 @@ func (h *AdminContactHandler) RejectInquiry(c *gin.Context) {
 		return
 	}
 
-	err = h.service.RejectInquiry(id)
+	err = h.service.RejectInquiry(int32(id))
 	if err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
