+++ backend/admin/repository/report_repository.go
@@ -17,7 +17,7 @@ func NewReportRepository(db *gorm.DB) *ReportRepository {
 }
 
 // FindAll retrieves all reports with pagination and optional filter
-func (r *ReportRepository) FindAll(page, pageSize int, handled *bool) ([]models.Report, int64, error) {
+func (r *ReportRepository) FindAll(page, pageSize int, handled *bool) ([]models.Report, int, error) {
 	var reports []models.Report
 	var total int64
 
@@ -34,11 +34,11 @@ func (r *ReportRepository) FindAll(page, pageSize int, handled *bool) ([]models.
 		return nil, 0, result.Error
 	}
 
-	return reports, total, nil
+	return reports, int(total), nil
 }
 
 // FindByID finds a report by ID
-func (r *ReportRepository) FindByID(id int) (*models.Report, error) {
+func (r *ReportRepository) FindByID(id int32) (*models.Report, error) {
 	var report models.Report
 	result := r.db.Where("reportId = ?", id).First(&report)
 	if result.Error != nil {
@@ -48,14 +48,14 @@ func (r *ReportRepository) FindByID(id int) (*models.Report, error) {
 }
 
 // CountUnprocessed counts unprocessed reports
-func (r *ReportRepository) CountUnprocessed() (int64, error) {
+func (r *ReportRepository) CountUnprocessed() (int, error) {
 	var count int64
 	result := r.db.Model(&models.Report{}).Where("reportFlag = ?", false).Count(&count)
-	return count, result.Error
+	return int(count), result.Error
 }
 
 // MarkAsHandled marks a report as handled
-func (r *ReportRepository) MarkAsHandled(id int) error {
+func (r *ReportRepository) MarkAsHandled(id int32) error {
 	return r.db.Model(&models.Report{}).
 		Where("reportId = ?", id).
 		Update("reportFlag", true).Error
