+++ backend/admin/service/admin_dashboard_service_test.go
@@ -14,14 +14,14 @@ type MockUserRepository struct {
 	mock.Mock
 }
 
-func (m *MockUserRepository) CountAll() (int64, error) {
+func (m *MockUserRepository) CountAll() (int, error) {
 	args := m.Called()
-	return args.Get(0).(int64), args.Error(1)
+	return args.Get(0).(int), args.Error(1)
 }
 
-func (m *MockUserRepository) CountByRole(role models.Role) (int64, error) {
+func (m *MockUserRepository) CountByRole(role models.Role) (int, error) {
 	args := m.Called(role)
-	return args.Get(0).(int64), args.Error(1)
+	return args.Get(0).(int), args.Error(1)
 }
 
 // MockPostRepository is a mock implementation of PostRepository
@@ -29,14 +29,14 @@ type MockPostRepository struct {
 	mock.Mock
 }
 
-func (m *MockPostRepository) CountAll() (int64, error) {
+func (m *MockPostRepository) CountAll() (int, error) {
 	args := m.Called()
-	return args.Get(0).(int64), args.Error(1)
+	return args.Get(0).(int), args.Error(1)
 }
 
-func (m *MockPostRepository) SumReactions() (int64, error) {
+func (m *MockPostRepository) SumReactions() (int, error) {
 	args := m.Called()
-	return args.Get(0).(int64), args.Error(1)
+	return args.Get(0).(int), args.Error(1)
 }
 
 // MockReportRepository is a mock implementation of ReportRepository
@@ -44,9 +44,9 @@ type MockReportRepository struct {
 	mock.Mock
 }
 
-func (m *MockReportRepository) CountUnprocessed() (int64, error) {
+func (m *MockReportRepository) CountUnprocessed() (int, error) {
 	args := m.Called()
-	return args.Get(0).(int64), args.Error(1)
+	return args.Get(0).(int), args.Error(1)
 }
 
 // MockBusinessMemberRepository is a mock implementation of BusinessMemberRepository
@@ -54,9 +54,9 @@ type MockBusinessMemberRepository struct {
 	mock.Mock
 }
 
-func (m *MockBusinessMemberRepository) CountAll() (int64, error) {
+func (m *MockBusinessMemberRepository) CountAll() (int, error) {
 	args := m.Called()
-	return args.Get(0).(int64), args.Error(1)
+	return args.Get(0).(int), args.Error(1)
 }
 
 func TestAdminDashboardService_GetSummary(t *testing.T) {
@@ -72,11 +72,11 @@ func TestAdminDashboardService_GetSummary(t *testing.T) {
 			UnprocessedReportCount: 5,
 		}
 
-		assert.Equal(t, int64(100), summary.TotalUserCount)
-		assert.Equal(t, int64(80), summary.ActiveUserCount)
-		assert.Equal(t, int64(500), summary.TotalPostCount)
-		assert.Equal(t, int64(1500), summary.TotalReactionCount)
-		assert.Equal(t, int64(25), summary.BusinessAccountCount)
-		assert.Equal(t, int64(5), summary.UnprocessedReportCount)
+		assert.Equal(t, 100, summary.TotalUserCount)
+		assert.Equal(t, 80, summary.ActiveUserCount)
+		assert.Equal(t, 500, summary.TotalPostCount)
+		assert.Equal(t, 1500, summary.TotalReactionCount)
+		assert.Equal(t, 25, summary.BusinessAccountCount)
+		assert.Equal(t, 5, summary.UnprocessedReportCount)
 	})
 }
