+++ backend/admin/service/admin_report_service.go
@@ -13,7 +13,7 @@ import (
 // ReportListResponse represents the paginated report list response
 type ReportListResponse struct {
 	Reports  []models.Report `json:"reports"`
-	Total    int64           `json:"total"`
+	Total    int             `json:"total"`
 	Page     int             `json:"page"`
 	PageSize int             `json:"pageSize"`
 }
@@ -77,16 +77,16 @@ func (s *AdminReportService) GetReports(page, pageSize int, handled *bool) (*Rep
 }
 
 // GetReportDetail retrieves a report with target post information
-func (s *AdminReportService) GetReportDetail(reportID int) (*ReportDetailResponse, error) {
+func (s *AdminReportService) GetReportDetail(reportID int32) (*ReportDetailResponse, error) {
 	report, err := s.reportRepo.FindByID(reportID)
 	if err != nil {
 		return nil, errors.New("report not found")
 	}
 
 	response := &ReportDetailResponse{
-		ReportID:     report.ReportID,
+		ReportID:     int(report.ReportID),
 		ReporterID:   report.UserID,
-		TargetPostID: report.PostID,
+		TargetPostID: int(report.PostID),
 		Reason:       report.Reason,
 		ReportedAt:   report.Date.Format("2006-01-02T15:04:05Z07:00"),
 		Handled:      report.ReportFlag,
@@ -97,7 +97,7 @@ func (s *AdminReportService) GetReportDetail(reportID int) (*ReportDetailRespons
 	var post models.Post
 	if err := s.db.Where("postId = ?", report.PostID).First(&post).Error; err == nil {
 		response.Post = &PostInfo{
-			PostID:   post.PostID,
+			PostID:   int(post.PostID),
 			Title:    post.Title,
 			Text:     post.Text,
 			UserID:   post.UserID,
@@ -112,7 +112,7 @@ func (s *AdminReportService) GetReportDetail(reportID int) (*ReportDetailRespons
 }
 
 // MarkAsHandled marks a report as handled
-func (s *AdminReportService) MarkAsHandled(id int) error {
+func (s *AdminReportService) MarkAsHandled(id int32) error {
 	// Verify the report exists
 	report, err := s.reportRepo.FindByID(id)
 	if err != nil {
